// src/graphql/generated/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /**
   * The `Date` scalar type represents a Date
   * value as specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  Date: { input: any; output: any; }
  /**
   * The `DateTime` scalar type represents a DateTime
   * value as specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  DateTime: { input: any; output: any; }
  /**
   * The `GenericScalar` scalar type represents a generic
   * GraphQL scalar value that could be:
   * String, Boolean, Int, Float, List or Object.
   */
  GenericScalar: { input: any; output: any; }
  /**
   * Allows use of a JSON String for input / output from the GraphQL schema.
   *
   * Use of this type is *not recommended* as you lose the benefits of having a defined, static
   * schema (one of the key benefits of GraphQL).
   */
  JSONString: { input: any; output: any; }
  /**
   * The `Time` scalar type represents a Time value as
   * specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  Time: { input: any; output: any; }
  /**
   * Create scalar that ignores normal serialization/deserialization, since
   * that will be handled by the multipart request spec
   */
  Upload: { input: any; output: any; }
};
// ...
export type AdditionalTestsNode = Node & {
  __typename?: 'AdditionalTestsNode';
  abrasionResistanceIIITest?: Maybe<AbrasionResistanceIiiTestNode>;
  abrasionResistanceIITest?: Maybe<AbrasionResistanceIiTestNode>;
  abrasionResistanceITest?: Maybe<AbrasionResistanceITestNode>;
  abrasionResistanceIVTest?: Maybe<AbrasionResistanceIvTestNode>;
  acceleratedAgeingTest?: Maybe<AcceleratedAgeingTestNode>;
  acceleratedLightAgingBendingTest?: Maybe<AcceleratedLightAgingBendingTestNode>;
  acceleratedLightAgingTearingStrengthTest?: Maybe<AcceleratedLightAgingTearingStrengthTestNode>;
  acceleratedLightAgingTest?: Maybe<AcceleratedLightAgingTestNode>;
  adhesionTest?: Maybe<AdhesionTestNode>;
  blockingTest?: Maybe<BlockingTestNode>;
  castorChairContinuousUseTest?: Maybe<CastorChairContinuousUseTestNode>;
  coldCrackTest?: Maybe<ColdCrackTestNode>;
  crockingResistanceIITest?: Maybe<CrockingResistanceIiTestNode>;
  crockingResistanceITest?: Maybe<CrockingResistanceITestNode>;
  dimensionalStabilityHeatTest?: Maybe<DimensionalStabilityHeatTestNode>;
  dimensionalStabilityHumidityTest?: Maybe<DimensionalStabilityHumidityTestNode>;
  dimensionalStabilityWaterTest?: Maybe<DimensionalStabilityWaterTestNode>;
  elongationAtBreakTest?: Maybe<ElongationAtBreakTestNode>;
  elongationIITest?: Maybe<ElongationIiTestNode>;
  elongationITest?: Maybe<ElongationITestNode>;
  flexResistanceIiTest?: Maybe<FlexResistanceIiTestNode>;
  flexResistanceTest?: Maybe<FlexResistanceTestNode>;
  flooringClassificationTest?: Maybe<FlooringClassificationTestNode>;
  hardnessTest?: Maybe<HardnessTestNode>;
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  lossOnHeatingTest?: Maybe<LossOnHeatingTestNode>;
  odourTest?: Maybe<OdourTestNode>;
  pattern: PatternNode;
  /** This file must be a PDF */
  pdf: Scalars['String']['output'];
  residualIndentationTest?: Maybe<ResidualIndentationTestNode>;
  resistanceToChemicalsTest?: Maybe<ResistanceToChemicalsTestNode>;
  seamFatigueTest?: Maybe<SeamFatigueTestNode>;
  seamStrengthTest?: Maybe<SeamStrengthTestNode>;
  setTest?: Maybe<SetTestNode>;
  slipResistanceOfFlooringTest?: Maybe<SlipResistanceOfFlooringTestNode>;
  stitchingStrengthTest?: Maybe<StitchingStrengthTestNode>;
  stretchTest?: Maybe<StretchTestNode>;
  tearStrengthTest?: Maybe<TearStrengthTestNode>;
  tearStrengthTongueMethodTest?: Maybe<TearStrengthTongueMethodTestNode>;
  tearStrengthTrapezoidMethodTest?: Maybe<TearStrengthTrapezoidMethodTestNode>;
  tensileStrengthTest?: Maybe<TensileStrengthTestNode>;
  thermalConductivityTest?: Maybe<ThermalConductivityTestNode>;
  thicknessOfWearLayerTest?: Maybe<ThicknessOfWearLayerTestNode>;
  volatilityTest?: Maybe<VolatilityTestNode>;
};
// ...
export type BackingNode = Node & {
  __typename?: 'BackingNode';
  description?: Maybe<Scalars['String']['output']>;
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  pattern: PatternNode;
  pinholes?: Maybe<Scalars['String']['output']>;
  pinholesMethod?: Maybe<Scalars['String']['output']>;
  pinholesResult?: Maybe<Scalars['String']['output']>;
  type?: Maybe<Scalars['String']['output']>;
  weight?: Maybe<Scalars['String']['output']>;
  weightMethod?: Maybe<GeneralWeightMethod>;
  weightResult?: Maybe<Scalars['String']['output']>;
  weightUnit?: Maybe<GeneralWeightUnit>;
};
// ...
export type BrandNodeConnection = {
  __typename?: 'BrandNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<BrandNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type CardOfColorsNode = Node & {
  __typename?: 'CardOfColorsNode';
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  pattern?: Maybe<PatternNode>;
  text: Scalars['String']['output'];
  textCa?: Maybe<Scalars['String']['output']>;
  textDe?: Maybe<Scalars['String']['output']>;
  textEn?: Maybe<Scalars['String']['output']>;
  textEs?: Maybe<Scalars['String']['output']>;
  textFr?: Maybe<Scalars['String']['output']>;
  textIt?: Maybe<Scalars['String']['output']>;
  textPl?: Maybe<Scalars['String']['output']>;
  textUk?: Maybe<Scalars['String']['output']>;
};
// ...
export type CatalogueNodeConnection = {
  __typename?: 'CatalogueNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<CatalogueNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type ContactNodeConnection = {
  __typename?: 'ContactNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<ContactNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type DownloadLogNodeConnection = {
  __typename?: 'DownloadLogNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<DownloadLogNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type EpdNodeConnection = {
  __typename?: 'EpdNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<EpdNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type InspirationNode = Node & {
  __typename?: 'InspirationNode';
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  imageThumbMedium?: Maybe<Scalars['String']['output']>;
  img: Scalars['String']['output'];
  pattern?: Maybe<PatternNode>;
  text: Scalars['String']['output'];
  textCa?: Maybe<Scalars['String']['output']>;
  textDe?: Maybe<Scalars['String']['output']>;
  textEn?: Maybe<Scalars['String']['output']>;
  textEs?: Maybe<Scalars['String']['output']>;
  textFr?: Maybe<Scalars['String']['output']>;
  textIt?: Maybe<Scalars['String']['output']>;
  textPl?: Maybe<Scalars['String']['output']>;
  textUk?: Maybe<Scalars['String']['output']>;
};
// ...
export type ManufacturingCertificationsNode = {
  __typename?: 'ManufacturingCertificationsNode';
  id: Scalars['ID']['output'];
  isSustainable: Scalars['Boolean']['output'];
  logo?: Maybe<Scalars['String']['output']>;
  manufacturingCertifications: PatternNodeConnection;
  name: Scalars['String']['output'];
};
// ...
export type MemoNode = Node & {
  __typename?: 'MemoNode';
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  size: Scalars['String']['output'];
};
// ...
export type Node = {
  /** The ID of the object. */
  id: Scalars['ID']['output'];
};
// ...
export enum PatternBusiness {
  /** N/A */
  A = 'A_',
  /** CADENA */
  Cadena = 'CADENA',
  /** COLOMBIA */
  Colombia = 'COLOMBIA',
  /** EUROPE */
  Europe = 'EUROPE',
  /** LATAM */
  Latam = 'LATAM',
  /** MEXICO */
  Mexico = 'MEXICO',
  /** USA */
  Usa = 'USA'
}
// ...
export type PatternNode = Node & {
  __typename?: 'PatternNode';
  /** Used to identify if has all required data to generate product overview */
  activeForProductOverview: Scalars['Boolean']['output'];
  additionalTests?: Maybe<Scalars['GenericScalar']['output']>;
  additionaltests?: Maybe<AdditionalTestsNode>;
  altName: Scalars['String']['output'];
  approvedCleanersPdfUrl?: Maybe<Scalars['String']['output']>;
  backing?: Maybe<BackingNode>;
  brands: BrandNodeConnection;
  business: PatternBusiness;
  cardOfColor?: Maybe<CardOfColorsNode>;
  catalogStockNumber?: Maybe<Scalars['String']['output']>;
  catalogues: CatalogueNodeConnection;
  /** Used to identify if Spradling is currently making catalogues of this Pattern */
  cataloguesAvailable: Scalars['Boolean']['output'];
  code: Scalars['String']['output'];
  contacts: ContactNodeConnection;
  created: Scalars['DateTime']['output'];
  date?: Maybe<Scalars['Date']['output']>;
  downloadLogs: DownloadLogNodeConnection;
  enNotes?: Maybe<Scalars['String']['output']>;
  epds: EpdNodeConnection;
  esNotes?: Maybe<Scalars['String']['output']>;
  general?: Maybe<GeneralNode>;
  /** Used to identify if there are available catalogues of this Pattern */
  hasStock: Scalars['Boolean']['output'];
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  imageThumbLarge?: Maybe<Scalars['String']['output']>;
  imageThumbMedium?: Maybe<Scalars['String']['output']>;
  imageThumbSmall?: Maybe<Scalars['String']['output']>;
  inspiration?: Maybe<InspirationNode>;
  isEpd?: Maybe<Scalars['Boolean']['output']>;
  /** Used to identify if this pattern is new */
  isNew: Scalars['Boolean']['output'];
  isVisibleLibrary?: Maybe<Scalars['Boolean']['output']>;
  maintenanceGuide?: Maybe<Scalars['String']['output']>;
  manufacturingCertifications?: Maybe<Array<Maybe<ManufacturingCertificationsNode>>>;
  memo?: Maybe<MemoNode>;
  modified: Scalars['DateTime']['output'];
  name: Scalars['String']['output'];
  nameWithSuperscriptR?: Maybe<Scalars['String']['output']>;
  observation?: Maybe<Scalars['String']['output']>;
  oraclerequestSet: OracleRequestNodeConnection;
  otherattributes?: Maybe<OtherAttributesNode>;
  /** Used to identify if this pattern is out of stock in Europe instance */
  outOfStockEurope: Scalars['Boolean']['output'];
  packagingFsc?: Maybe<Scalars['Boolean']['output']>;
  patents: PatentNodeConnection;
  pattern: ServiceContactNodeConnection;
  patternClass?: Maybe<Scalars['String']['output']>;
  patterns: PromotedProjectNodeConnection;
  pdfUrl?: Maybe<Scalars['String']['output']>;
  productcategorization?: Maybe<ProductCategorizationNode>;
  products?: Maybe<ProductNodeConnection>;
  projects: ProjectNodeConnection;
  promotedProjects?: Maybe<Array<Maybe<PromotedProjectNode>>>;
  quotations: QuotationNodeConnection;
  relatedPatterns?: Maybe<PatternNodeConnection>;
  rev?: Maybe<Scalars['Int']['output']>;
  saleable?: Maybe<Scalars['Boolean']['output']>;
  samplescontact: SampleContactNodeConnection;
  sectors: SectorNodeConnection;
  sliders?: Maybe<Scalars['GenericScalar']['output']>;
  slug: Scalars['String']['output'];
  specsSheetData?: Maybe<Scalars['GenericScalar']['output']>;
  status?: Maybe<Scalars['String']['output']>;
  stores: StoreNodeConnection;
  successFlameRetardancyTests?: Maybe<Scalars['GenericScalar']['output']>;
  successPdfFlameRetardancyTests?: Maybe<Scalars['GenericScalar']['output']>;
  sustainability?: Maybe<SustainabilityNode>;
  transformationGuide?: Maybe<Scalars['String']['output']>;
  trend?: Maybe<TrendNode>;
  wishlistItems: WishlistItemNodeConnection;
  zippedImagesUrl?: Maybe<Scalars['String']['output']>;
};